Refactor: UI에 사용되는 데이터 모델을 ui:data 모듈로 이동
