+++ ui/creategroup/build.gradle.kts
@@ -4,4 +4,5 @@ android {
 
 dependencies {
     implementation(project(":ui:expenselist"))
+    implementation(project(":ui:data"))
 }
