--- ui/creategroup/src/main/java/com/kappzzang/jeongsan/creategroup/Member.kt
@@ -1,3 +1,0 @@
-package com.kappzzang.jeongsan.creategroup
-
-class Member(val name: String)
