+++ ui/creategroup/src/main/java/com/kappzzang/jeongsan/creategroup/MemberAdapter.kt
@@ -5,6 +5,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.kappzzang.jeongsan.data.Member
 import com.kappzzang.jeongsan.expenselist.R
 
 class MemberAdapter(var memberList: List<Member>, val inflater: LayoutInflater, val layoutId: Int) :
