+++ ui/data/src/main/java/com/kappzzang/jeongsan/data/GroupViewItem.kt
@@ -1,4 +1,4 @@
-package com.kappzzang.jeongsan.main
+package com.kappzzang.jeongsan.data
 
 import com.kappzzang.jeongsan.model.GroupItem
