+++ ui/data/src/main/java/com/kappzzang/jeongsan/data/Member.kt
@@ -1,0 +1,3 @@
+package com.kappzzang.jeongsan.data
+
+class Member(val name: String)
