+++ ui/main/build.gradle.kts
@@ -7,4 +7,5 @@ dependencies {
     implementation(project(":ui:creategroup"))
     implementation(project(":domain:group"))
     implementation(project(":domain:user"))
+    implementation(project(":ui:data"))
 }
